Replace gui_str_pool with shared u16 name pool
Inline the per-registry gui_str_pool_t generic primitive into a single shared implementation in gui_names.c, narrowing all name_off fields and the intern/cstr API from u32 to u16. The u16 offset caps the pool at 64 KB; a GUI_NAMES_NONE sentinel (0xffff) replaces GUI_STR_POOL_NONE. Delete gui_str_pool.h now that it has no other users.

+++ source/runtime_service/gui/draw/gui_icon.c
@@ -80,7 +80,7 @@
 
 typedef struct
 {
-    u32  name_off;  // lookup key -- offset into the shared name pool (gui_names.h)
+    u16  name_off;  // lookup key -- offset into the shared name pool (gui_names.h)
     u16  w, h;      // STORED pixel dimensions (the field's, for an SDF icon)
     u16  tenant;    // packed: bits [0,15) = atlas handle (0 = unused), bit 15 = sdf flag
